fix: decompose func
